+++ src/azure/core/lease.rs
@@ -2,6 +2,36 @@ use std::str::FromStr;
 use std::string::ParseError;
 use std::fmt::{Display, Formatter};
 use std::fmt;
+use azure::core::enumerations;
+use azure::core::errors::TraversingError;
+use azure::core::parsing::FromStringOptional;
+
+create_enum!(LeaseStatus,
+                            (Locked,        "locked"),
+                            (Unlocked,      "unlocked")
+);
+
+create_enum!(LeaseState,
+                            (Available,     "available"),
+                            (Leased,        "leased"),
+                            (Expired,       "expired"),
+                            (Breaking,      "breaking"),
+                            (Broken,        "broken")
+);
+
+create_enum!(LeaseDuration,
+                            (Infinite,      "infinite"),
+                            (Fixed,         "fixed")
+);
+
+create_enum!(LeaseAction,
+                            (Acquire,      "acquire"),
+                            (Renew,         "renew "),
+                            (Change,        "change"),
+                            (Release,        "release "),
+                            (Break,         "break")
+);
+
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct LeaseId {